# specs/001-rag-chatbot-auth/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Physical AI & Humanoid Robotics RAG Chatbot API
  description: Backend API for authenticated RAG chatbot with conversation history, recommendations, and reading progress tracking
  version: 1.0.0
  contact:
    name: API Support
    email: support@physical-ai-book.com
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.physical-ai-book.com
    description: Production server (Hugging Face Spaces)
...
security:
  - bearerAuth: []
paths:
...
  /auth/signup:
    post:
      tags:
        - Authentication
      summary: Create new user account
      description: Register a new user with email, password, and background questionnaire
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - full_name
                - experience_level
                - programming_languages
                - frameworks
                - available_hardware
                - robotics_hardware
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  minLength: 8
                  example: SecurePass123
                full_name:
                  type: string
                  minLength: 2
                  example: John Doe
                experience_level:
                  type: string
                  enum: [Beginner, Intermediate, Advanced, Expert]
                  example: Intermediate
                programming_languages:
                  type: array
                  items:
                    type: string
                    enum: [Python, "C++", JavaScript, Rust, Go, Java, MATLAB, Julia, Other]
                  minItems: 1
                  example: [Python, "C++"]
                frameworks:
                  type: array
                  items:
                    type: string
                    enum: [ROS, "ROS 2", PyTorch, TensorFlow, Unity, "Unreal Engine", Gazebo, Mujoco, "Isaac Sim", Other]
                  example: ["ROS 2", PyTorch]
                available_hardware:
                  type: array
                  items:
                    type: string
                    enum: ["CPU only", "NVIDIA GPU", "AMD GPU", "Apple Silicon (M1/M2)", "NVIDIA Jetson", "Google TPU", Other]
                  example: ["NVIDIA GPU"]
                robotics_hardware:
                  type: array
                  items:
                    type: string
                    enum: ["None (simulation only)", "Educational robot kit", "Research robotic arm", "Mobile robot", "Humanoid robot", "Custom hardware", Other]
                  example: ["None (simulation only)"]
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid input (validation errors)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (max 5 attempts per minute)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signin:
    post:
      tags:
        - Authentication
      summary: Sign in existing user
      description: Authenticate user with email and password, returns session token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  example: SecurePass123
                remember_me:
                  type: boolean
                  default: false
                  description: Extend session expiration to 7 days
      responses:
        '200':
          description: Signin successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (max 5 attempts per minute)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/signout:
    post:
      tags:
        - Authentication
      summary: Sign out current user
      description: Invalidate current session token
      responses:
        '200':
          description: Signout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Signout successful
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /auth/me:
    get:
      tags:
        - Authentication
      summary: Get current user profile
      description: Retrieve authenticated user's profile and background information
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

  # Chat Endpoints
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
  responses:
    UnauthorizedError:
      description: Authentication required or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    AuthResponse:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/UserProfile'
        token:
          type: string
          description: JWT session token (also set as HTTP-only cookie)
        expires_at:
          type: string
          format: date-time
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        experience_level:
          type: string
          enum: [Beginner, Intermediate, Advanced, Expert]
        programming_languages:
          type: array
          items:
            type: string
        frameworks:
          type: array
          items:
            type: string
        available_hardware:
          type: array
          items:
            type: string
        robotics_hardware:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...
    Error:
      type: object
      properties:
        error:
          type: string
        details:
          type: object
          additionalProperties: true